+++ src/server/core/storage/storage.go
@@ -8,7 +8,7 @@ import (
 	"github.com/ai-microsoft/haystack/server/core/pebble"
 )
 
-const StorageVersion = "1.2"
+const StorageVersion = "1.3"
 
 func cleanup(storagePath string) {
 	// Perform cleanup tasks here, such as removing old files or directories
@@ -17,6 +17,7 @@ func cleanup(storagePath string) {
 		"index", // It's the first version of the index, we can safely remove it now
 		"1.0",
 		"1.1",
+		"1.2",
 	}
 
 	for _, item := range cleanupList {